gmeredith_CS446_PA4/requirements.txt
Pillow
SciPy
numpy
pyOpenSSL